Extract getPostAndMorePosts
