+++ examples/cms-datocms/lib/api.js
@@ -4,7 +4,7 @@ const API_URL = 'https://graphql.datocms.com'
 const API_TOKEN = process.env.NEXT_EXAMPLE_CMS_DATOCMS_API_TOKEN
 
 // See: https://www.datocms.com/blog/offer-responsive-progressive-lqip-images-in-2020
-export const responsiveImageFragment = `
+const responsiveImageFragment = `
   fragment responsiveImageFragment on ResponsiveImage {
   srcSet
     webpSrcSet
@@ -20,7 +20,7 @@ export const responsiveImageFragment = `
   }
 `
 
-export default async function fetchAPI(query, { variables, preview } = {}) {
+async function fetchAPI(query, { variables, preview } = {}) {
   const res = await fetch(API_URL + (preview ? '/preview' : ''), {
     method: 'POST',
     headers: {
@@ -98,4 +98,60 @@ export async function getAllPostsForHome(preview) {
     { preview }
   )
   return data?.allPosts
+}
+
+export async function getPostAndMorePosts(slug, preview) {
+  const data = await fetchAPI(
+    `
+  query PostBySlug($slug: String) {
+    post(filter: {slug: {eq: $slug}}) {
+      title
+      slug
+      content
+      date
+      ogImage: coverImage{
+        url(imgixParams: {fm: jpg, fit: crop, w: 2000, h: 1000 })
+      }
+      coverImage {
+        responsiveImage(imgixParams: {fm: jpg, fit: crop, w: 2000, h: 1000 }) {
+          ...responsiveImageFragment
+        }
+      }
+      author {
+        name
+        picture {
+          url(imgixParams: {w: 100, h: 100, sat: -100})
+        }
+      }
+    }
+
+    morePosts: allPosts(orderBy: date_DESC, first: 2, filter: {slug: {neq: $slug}}) {
+      title
+      slug
+      excerpt
+      date
+      coverImage {
+        responsiveImage(imgixParams: {fm: jpg, fit: crop, w: 2000, h: 1000 }) {
+          ...responsiveImageFragment
+        }
+      }
+      author {
+        name
+        picture {
+          url(imgixParams: {w: 100, h: 100, sat: -100})
+        }
+      }
+    }
+  }
+
+  ${responsiveImageFragment}
+  `,
+    {
+      preview,
+      variables: {
+        slug,
+      },
+    }
+  )
+  return data
 }
